+++ models/FileSystemItem.php
@@ -2,6 +2,7 @@
 
 namespace humhub\modules\cfiles\models;
 
+use humhub\modules\cfiles\Module;
 use humhub\modules\cfiles\permissions\ManageFiles;
 use humhub\modules\content\components\ContentContainerActiveRecord;
 use humhub\modules\content\components\ContentActiveRecord;
@@ -27,6 +28,11 @@ abstract class FileSystemItem extends ContentActiveRecord implements ItemInterfa
     public $visibility;
 
     /**
+     * @var ?int used for edit form
+     */
+    public $hidden = null;
+
+    /**
      * @inheritdoc
      */
     public $managePermission = ManageFiles::class;
@@ -47,15 +53,21 @@ abstract class FileSystemItem extends ContentActiveRecord implements ItemInterfa
     public function rules()
     {
         return [
-            ['visibility', 'integer', 'min' => 0, 'max' => 1]
+            ['visibility', 'integer', 'min' => 0, 'max' => 1],
+            ['hidden', 'boolean']
         ];
     }
 
     abstract function updateVisibility($visibility);
+
     abstract function getSize();
+
     abstract function getItemType();
+
     abstract function getDescription();
+
     abstract function getDownloadCount();
+
     abstract function getVisibilityTitle();
 
     /**
@@ -75,6 +87,7 @@ abstract class FileSystemItem extends ContentActiveRecord implements ItemInterfa
     {
         return [
             'visibility' => Yii::t('CfilesModule.models_FileSystemItem', 'Is Public'),
+            'hidden' => Yii::t('CfilesModule.models_FileSystemItem', 'Hide in Stream'),
             'download_count' => Yii::t('CfilesModule.models_FileSystemItem', 'Downloads'),
         ];
     }
@@ -82,22 +95,12 @@ abstract class FileSystemItem extends ContentActiveRecord implements ItemInterfa
     /**
      * @inheritdoc
      */
-    public function afterFind() {
+    public function afterFind()
+    {
         $this->visibility = $this->content->visibility;
+        $this->hidden = $this->content->hidden;
         parent::afterFind();
     }
-
-    public function handleContentSave($evt, $content = null)
-    {
-        /* @var $content Content */
-        $content = ($content) ? $content : $evt->sender;
-        if($evt->sender->container instanceof User && $evt->sender->isPrivate()) {
-            $evt->sender->visibility = Content::VISIBILITY_OWNER;
-        }
-
-        return true;
-    }
-
 
     /**
      * @inheritdoc
@@ -106,6 +109,12 @@ abstract class FileSystemItem extends ContentActiveRecord implements ItemInterfa
     {
         if ($this->parent_folder_id == "") {
             $this->parent_folder_id = null;
+        }
+
+        if ($insert && $this->hidden === null) {
+            /** @var Module $module */
+            $module = Yii::$app->getModule('cfiles');
+            $this->hidden = $module->getContentHiddenDefault($this->content->container);
         }
 
         return parent::beforeSave($insert);
@@ -119,13 +128,19 @@ abstract class FileSystemItem extends ContentActiveRecord implements ItemInterfa
         // this should set the editor and edit date of all parent folders if sth. inside of them has changed
         if (!empty($this->parentFolder)) {
             $this->parentFolder->save();
-            if($this->parentFolder->content->isPrivate() && $this->content->isPublic()) {
+            if ($this->parentFolder->content->isPrivate() && $this->content->isPublic()) {
                 $this->content->visibility = Content::VISIBILITY_PRIVATE;
                 $this->content->save();
             }
         }
 
+        $this->content->hidden = $this->hidden;
+        if (!$insert) {
+            $this->content->save();
+        }
+
         parent::afterSave($insert, $changedAttributes);
+
     }
 
     /**