+++ views/edit/modal_edit_file.php
@@ -3,6 +3,7 @@
 
 use humhub\modules\content\widgets\richtext\RichTextField;
 use humhub\modules\topic\widgets\TopicPicker;
+use humhub\modules\ui\form\widgets\ContentHiddenCheckbox;
 use humhub\modules\ui\form\widgets\ContentVisibilitySelect;
 use humhub\widgets\ModalButton;
 use humhub\widgets\ModalDialog;
@@ -25,6 +26,7 @@ use yii\bootstrap\ActiveForm;
             <?= $form->field($file, 'description')->widget(RichTextField::class); ?>
             <?= $form->field($file, 'topics')->widget(TopicPicker::class, ['contentContainer' => $file->content->container])->label(false); ?>
             <?= $form->field($file, 'visibility')->widget(ContentVisibilitySelect::class, ['readonly' => $file->parentFolder->content->isPrivate()]); ?>
+            <?= $form->field($file, 'hidden')->widget(ContentHiddenCheckbox::class, []); ?>
             <?= $form->field($file, 'download_count')->staticControl(['style' => 'display:inline']); ?>
         </div>